backend/requirements.txt
fastapi
uvicorn
motor
python-dotenv
PyJWT
cloudinary
python-multipart
